backend-fastapi/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
sqlalchemy>=2.0
pydantic>=2.0
pydantic-settings>=2.0
polars>=1.0
python-multipart>=0.0
python-dotenv>=1.0
openpyxl>=3.0
supabase>=2.0
pandas>=2.0
requests>=2.0
